+++ gold-wallet-service/src/main/java/com/melli/wallet/service/WalletCashLimitationService.java
@@ -6,10 +6,12 @@ import com.melli.wallet.domain.master.entity.WalletEntity;
 import com.melli.wallet.domain.redis.WalletCashInLimitationRedis;
 import com.melli.wallet.exception.InternalServiceException;
 
+import java.math.BigDecimal;
+
 public interface WalletCashLimitationService {
 
-	void checkCashInLimitation(ChannelEntity channel, WalletEntity wallet, long amount, WalletAccountEntity walletAccount) throws InternalServiceException;
+	void checkCashInLimitation(ChannelEntity channel, WalletEntity wallet, BigDecimal amount, WalletAccountEntity walletAccount) throws InternalServiceException;
 
-	void updateCashInLimitation(WalletAccountEntity wallet, long amount) throws InternalServiceException;
+	void updateCashInLimitation(WalletAccountEntity wallet, BigDecimal amount) throws InternalServiceException;
 
 }
